+++ src/canister/user_index/src/api/canister_lifecycle/post_upgrade.rs
@@ -16,7 +16,6 @@ use crate::{
 fn post_upgrade() {
     restore_data_from_stable_memory();
     update_version_from_args();
-    reclaim_cycles_from_individual_canisters();
     insert_platform_orchestrator_in_well_known_principal();
 }
 
@@ -24,22 +23,6 @@ fn insert_platform_orchestrator_in_well_known_principal() {
     CANISTER_DATA.with_borrow_mut(|canister_data| {
         canister_data.configuration.known_principal_ids.insert(CanisterIdPlatformOrchestrator, Principal::from_text("74zq4-iqaaa-aaaam-ab53a-cai").unwrap());
     })
-}
-
-
-fn reclaim_cycles_from_individual_canisters() {
-    ic_cdk::spawn(impl_reclaim_cycles_from_individual_canisters_and_send_to_plaform_orchestrator())
-}
-
-async fn impl_reclaim_cycles_from_individual_canisters_and_send_to_plaform_orchestrator() {
-    let canister_ids = CANISTER_DATA.with_borrow(|canister_data| {
-        canister_data.user_principal_id_to_canister_id_map.clone().into_values()
-    });
-
-    let relcaim_cycles_from_canister_futures = canister_ids.map(|canister_id| {
-        call::<_ , ()>(canister_id, "return_cycles_to_user_index_canister", ())
-    });
-    run_task_concurrently(relcaim_cycles_from_canister_futures, 10, |_| {}, || false).await;
 }
 
 fn update_version_from_args() {
