+++ src/canister/user_index/src/api/cycle_management/mod.rs
@@ -1,1 +1,2 @@
 pub mod get_user_index_canister_cycle_balance;
+pub mod reclaim_cycle_from_individual_canisters;
